Fix Compiler Warnings

+++ src/Dielectric.cpp
@@ -13,11 +13,11 @@ Dielectric::Dielectric(float refractionIndex)
 const bool Dielectric::Scatter(const Ray& ray, const HitRecord& record, glm::vec3& attenuation, Ray& scattered) const
 {
     attenuation = glm::vec3(1.0f);
-    float refractionRatio = record.frontFace ? (1.0 / m_RefractionIndex) : m_RefractionIndex;
+    float refractionRatio = record.frontFace ? (1.0f / m_RefractionIndex) : m_RefractionIndex;
 
     glm::vec3 unitDir = glm::normalize(ray.Direction);
-    double cosTheta = glm::min(glm::dot(-unitDir, record.Normal), 1.0f);
-    double sinTheta = glm::sqrt(1.0f - cosTheta * cosTheta);
+    float cosTheta = glm::min(glm::dot(-unitDir, record.Normal), 1.0f);
+    float sinTheta = glm::sqrt(1.0f - cosTheta * cosTheta);
 
     bool cannotRefract = refractionRatio * sinTheta > 1.0f;
     glm::vec3 direction;
